resolving PR comments
