--- external/config-checker/Common/configurations/jaggery-conf-ref.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<configurations>
-   <!-- <config id="HttpHeaderSecurityFilter">
-        <![CDATA[
-   {
-     "filters":[
-        {
-            "name":"HttpHeaderSecurityFilter",
-            "class":"org.apache.catalina.filters.HttpHeaderSecurityFilter",
-            "params" : [{"name" : "hstsEnabled"}]
-        }
-        ]
-   }
-]]>
-    </config>-->
-
-   <!-- <config id="HttpHeaderSecurityFilterMapping">
-        <![CDATA[
-   {
-     "filterMappings" :[
-    {
-    "name":"HttpHeaderSecurityFilter",
-    "url":"*"
-    }]
-   }
-]]>
-    </config>-->
-
-</configurations>
