--- external/config-checker/src/main/assembly/zip.xml
@@ -1,52 +1,0 @@
-<!--
- Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at
- http://www.apache.org/licenses/LICENSE-2.0
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
--->
-<assembly xmlns="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.2"
-          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-          xsi:schemaLocation="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.2 http://maven.apache.org/xsd/assembly-1.1.2.xsd">
-    <id>pack</id>
-    <includeBaseDirectory>true</includeBaseDirectory>
-
-    <formats>
-        <format>zip</format>
-    </formats>
-    <fileSets>
-        <fileSet>
-            <directory>${project.basedir}/Common</directory>
-        </fileSet>
-        <fileSet>
-            <directory>${project.basedir}/Product</directory>
-        </fileSet>
-    </fileSets>
-    <files>
-        <file>
-            <source> ${project.build.directory}/${project.artifactId}-${project.version}.jar</source>
-            <outputDirectory>/</outputDirectory>
-        </file>
-        <file>
-            <source>config-checker-settings.yaml</source>
-            <outputDirectory>/</outputDirectory>
-        </file>
-        <file>
-            <source>src/main/resources/log4j.properties</source>
-            <outputDirectory>/</outputDirectory>
-        </file>
-    </files>
-    <dependencySets>
-        <dependencySet>
-            <outputDirectory>/lib</outputDirectory>
-            <excludes>
-                <exclude>${project.groupId}-:${project.artifactId}:jar:*</exclude>
-            </excludes>
-        </dependencySet>
-    </dependencySets>
-</assembly>
