--- external/config-checker/src/main/java/org/wso2/security/tools/exception/ConfigCheckerException.java
@@ -1,15 +1,0 @@
-package org.wso2.security.tools.exception;
-
-/**
- * Exception wrapper class for Config Checker.
- */
-public class ConfigCheckerException extends Exception {
-
-    public ConfigCheckerException(String message, Throwable e) {
-        super(message);
-    }
-
-    public ConfigCheckerException(String message) {
-        super(message);
-    }
-}
