--- external/config-checker/src/main/java/org/wso2/security/tools/model/ChildProperty.java
@@ -1,71 +1,0 @@
-/*
- *  Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-package org.wso2.security.tools.model;
-
-import java.util.ArrayList;
-import java.util.List;
-
-/**
- * Model class for child properties.
- */
-public class ChildProperty {
-
-    private String refChildConfigFilePath = null;
-    private String productConfigFIlePath = null;
-    private String fileFormat = null;
-    private String childFileType = null;
-    private List<String> excludeList = new ArrayList<String>();
-
-    public List<String> getExcludeList() {
-        return excludeList;
-    }
-
-    public void setExcludeList(List<String> excludeList) {
-        this.excludeList = excludeList;
-    }
-
-    public String getRefChildConfigFilePath() {
-        return refChildConfigFilePath;
-    }
-
-    public void setRefChildConfigFilePath(String refChildConfigFilePath) {
-        this.refChildConfigFilePath = refChildConfigFilePath;
-    }
-
-    public String getProductConfigFilePath() {
-        return productConfigFIlePath;
-    }
-
-    public void setProductConfigFIlePath(String productConfigFIlePath) {
-        this.productConfigFIlePath = productConfigFIlePath;
-    }
-
-    public String getFileFormat() {
-        return fileFormat;
-    }
-
-    public void setFileFormat(String fileFormat) {
-        this.fileFormat = fileFormat;
-    }
-
-    public String getChildFileType() {
-        return childFileType;
-    }
-
-    public void setChildFileType(String childFileType) {
-        this.childFileType = childFileType;
-    }
-}
