--- external/config-checker/src/main/java/org/wso2/security/tools/model/Configuration.java
@@ -1,40 +1,0 @@
-/*
- *  Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-package org.wso2.security.tools.model;
-
-/**
- * Model class for reference configuration blocks.
- */
-public class Configuration {
-    private String configurationID;
-    private String configurationValue;
-
-    public String getConfigurationID() {
-        return configurationID;
-    }
-
-    public void setConfigurationID(String configurationID) {
-        this.configurationID = configurationID;
-    }
-
-    public String getConfigurationValue() {
-        return configurationValue;
-    }
-
-    public void setConfigurationValue(String configurationValue) {
-        this.configurationValue = configurationValue;
-    }
-}
