--- external/config-checker/src/main/java/org/wso2/security/tools/utils/Constants.java
@@ -1,39 +1,0 @@
-/*
- *  Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-package org.wso2.security.tools.utils;
-
-/**
- * A Class containing the utils constants.
- */
-public final class Constants {
-
-    public static final String FILE_TYPE_TAG = "type";
-    public static final String FILE_FORMAT_TAG = "format";
-    public static final String CONFIG_ID_ATTRIBUTE = "id";
-    public static final String REFERENCE_PATH_TAG = "reference-path";
-    public static final String PRODUCT_PATH_TAG = "product-path";
-    public static final String EXCLUDE_CONFIGS_TAG = "exclude-configs";
-    public static final String CONFIGURATIONS_TAG = "configurations";
-    public static final String EXCLUDE_PATH_TAG = "path";
-    public static final String EXCLUDE_PATHS_TAG = "exclude-paths";
-    public static final String PARENT_PROPERTIES_TAG = "parent-properties";
-    public static final String CHILD_PROPERTIES_TAG = "child-properties";
-    public static final String FILE_TAG = "file";
-    public static final String CONFIG_TAG = "config";
-
-    public static final String NEWLINES_REGEX = "[\\n]+";
-    public static final String PRESENT_DIRECTORY = ".";
-}
