+++ internal/config-checker/common/properties/parent.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0"  ?>
+<!--
+  ~ Copyright (c) 2019, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+  ~
+  ~ WSO2 Inc. licenses this file to you under the Apache License,
+  ~ Version 2.0 (the "License"); you may not use this file except
+  ~ in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~    http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing,
+  ~ software distributed under the License is distributed on an
+  ~ "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+  ~ KIND, either express or implied.  See the License for the
+  ~ specific language governing permissions and limitations
+  ~ under the License.
+  -->
+
+<!--Parent properties file containing the reference common configuration file paths -->
+<parent-properties>
+    <!--<file type="web.xml">
+        <format>xml</format>
+        <reference-path>/common/configurations/web-xml-ref.xml</reference-path>
+    </file>-->
+    <!--<file type="jaggery.conf">
+        <format>json</format>
+        <reference-path>/common/configurations/jaggery-conf-ref.xml</reference-path>
+    </file>-->
+</parent-properties>
