+++ internal/config-checker/config-checker-settings.yaml
@@ -1,0 +1,27 @@
+#
+# Copyright 2019 WSO2, Inc. (http://wso2.com)
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+# Config Checker settings
+
+parentPropertiesFile : common/properties/parent.xml
+childPropertiesFile : product/properties/child.xml
+
+parsers:
+ xml : org.wso2.security.tools.configchecker.parser.xml.XMLConfigurationParser
+ json : org.wso2.security.tools.configchecker.parser.json.JSONConfigurationParser
+
+reportGenerators:
+ text: org.wso2.security.tools.configchecker.report.TextReportGenerator
