+++ internal/config-checker/product/configurations/tomcat/web-xml-ref.xml
@@ -1,0 +1,39 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+  ~ Copyright (c) 2019, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+  ~
+  ~ WSO2 Inc. licenses this file to you under the Apache License,
+  ~ Version 2.0 (the "License"); you may not use this file except
+  ~ in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~    http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing,
+  ~ software distributed under the License is distributed on an
+  ~ "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+  ~ KIND, either express or implied.  See the License for the
+  ~ specific language governing permissions and limitations
+  ~ under the License.
+  -->
+
+<configurations>
+   <!-- <config id="HttpHeaderSecurityFilter">
+        <![CDATA[
+   <web-app>
+    <filter>
+ <filter-name>HttpHeaderSecurityFilter</filter-name>
+ <filter-class>org.apache.catalina.filters.HttpHeaderSecurityFilter</filter-class>
+ <init-param>
+ <param-name>hstsEnabled</param-name>
+        <param-value>false</param-value>
+      </init-param>
+      <init-param>
+        <param-name>antiClickJackingEnabled</param-name>
+        <param-value>false</param-value>
+      </init-param>
+    </filter>
+    </web-app>
+]]>
+    </config>-->
+</configurations>
