+++ internal/config-checker/src/main/java/org/wso2/security/tools/configchecker/model/ParentProperty.java
@@ -1,0 +1,41 @@
+/*
+ *  Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+package org.wso2.security.tools.configchecker.model;
+
+/**
+ * Model class for parent properties.
+ */
+public class ParentProperty {
+
+    private String refParentConfigFilePath = null;
+    private String fileFormat = null;
+
+    public String getRefParentConfigFilePath() {
+        return refParentConfigFilePath;
+    }
+
+    public void setRefParentConfigFilePath(String refParentConfigFilePath) {
+        this.refParentConfigFilePath = refParentConfigFilePath;
+    }
+
+    public String getFileFormat() {
+        return fileFormat;
+    }
+
+    public void setFileFormat(String fileFormat) {
+        this.fileFormat = fileFormat;
+    }
+}
